+++ utilcode/lib/src/main/java/com/blankj/utilcode/util/ResourceUtils.java
@@ -32,6 +32,16 @@ public final class ResourceUtils {
     }
 
     /**
+     * Return the drawable identifier by name.
+     *
+     * @param name The name of drawable.
+     * @return the drawable identifier by name
+     */
+    public static int getDrawableIdByName(String name) {
+        return Utils.getApp().getResources().getIdentifier(name, "drawable", Utils.getApp().getPackageName());
+    }
+
+    /**
      * Copy the file from assets.
      *
      * @param assetsFilePath The path of file in assets.