+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* `19/04/25` [fix] LogUtils delete due log.
 * `19/04/24` [upd] The swipe panel.
 * `19/03/17` [fix] The ugly UI.
 * `19/03/14` [fix] AdaptScreenUtils didn't work on some HaWei tablet.